:recycle:Refactor: split strategy component 선물&현물 페이지, 세부 페이지, 옵션 컴포넌트 분리
