+++ tradin/app/future/[num]/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import FutureDetail from "../../../src/components/units/strategy/future/FutureDetail";
+import FutureDetail from "../../../src/components/units/strategy/future/detail/FutureDetail.container";
 
 export default function FutureDetailPage(): JSX.Element {
     return (
