+++ tradin/app/future/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import FuturesPage from "../../src/components/units/strategy/future/Futures";
+import FuturesPage from "../../src/components/units/strategy/future/Future.container";
 
 export default function Futures() {
     return (
