+++ tradin/app/spot/[num]/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import SpotDetail from "../../../src/components/units/strategy/spot/SpotDetail";
+import SpotDetail from "../../../src/components/units/strategy/spot/detail/SpotDetail.container";
 
 export default function SpotDetailPage(): JSX.Element {
     return (
