+++ tradin/src/components/units/strategy/future/Future.container.tsx
@@ -1,0 +1,19 @@
+import { useRouter } from 'next/navigation';
+import FuturePageUI from "./Future.presenter";
+
+export default function FuturePage(): JSX.Element {
+    const router = useRouter();
+
+    const onClickMoveToFutureStrategy = (num: number) => {
+        const queryString = `currentStrategy=${num}`;
+        router.push(`./future/${num}?${queryString}`);
+    };
+
+    return (
+        <>
+            <FuturePageUI
+                onClickMoveToFutureStrategy={onClickMoveToFutureStrategy}
+            />
+        </>
+    );
+}
