+++ tradin/src/components/units/strategy/spot/Spot.types.tsx
@@ -1,0 +1,16 @@
+type Position = '상승' | '하강';
+
+export interface CoinData {
+    position: Position;
+    entryPrice: string;
+    profitLoss: string;
+    winRate: string;
+    profitFactor: string;
+    trades: number;
+    averageBars: number;
+    averageProfit: string;
+}
+
+export interface SpotPageUIProps {
+    onClickMoveToSpotStrategy: (num: number) => void;
+}
