+++ tradin/src/components/units/strategy/spot/detail/MockTransaction.tsx
@@ -1,0 +1,19 @@
+export const transactions = [
+    { id: 1, type: "BUY", date: "2024-06-01", price: "1000", yield: "10.00%", cumYield: "50.00%" },
+    { id: 1, type: "SELL", date: "2024-06-02", price: "1100", yield: "", cumYield: "" },
+    { id: 2, type: "BUY", date: "2024-06-03", price: "2000", yield: "20.00%", cumYield: "70.00%" },
+    { id: 2, type: "SELL", date: "2024-06-04", price: "2100", yield: "", cumYield: "" },
+    { id: 3, type: "BUY", date: "2024-06-03", price: "2000", yield: "20.00%", cumYield: "70.00%" },
+    { id: 3, type: "SELL", date: "2024-06-04", price: "2100", yield: "", cumYield: "" },
+    { id: 4, type: "BUY", date: "2024-06-03", price: "2000", yield: "20.00%", cumYield: "70.00%" },
+    { id: 4, type: "SELL", date: "2024-06-04", price: "2100", yield: "", cumYield: "" },
+    { id: 5, type: "BUY", date: "2024-06-03", price: "2000", yield: "20.00%", cumYield: "70.00%" },
+    { id: 5, type: "SELL", date: "2024-06-04", price: "2100", yield: "", cumYield: "" },
+    { id: 6, type: "BUY", date: "2024-06-03", price: "2000", yield: "20.00%", cumYield: "70.00%" },
+    { id: 6, type: "SELL", date: "2024-06-04", price: "2100", yield: "", cumYield: "" },
+    { id: 7, type: "BUY", date: "2024-06-03", price: "2000", yield: "20.00%", cumYield: "70.00%" },
+    { id: 7, type: "SELL", date: "2024-06-04", price: "2100", yield: "", cumYield: "" },
+    { id: 8, type: "BUY", date: "2024-06-03", price: "2000", yield: "20.00%", cumYield: "70.00%" },
+    { id: 8, type: "SELL", date: "2024-06-04", price: "2100", yield: "", cumYield: "" },
+    // Add more transactions as needed
+];
